Update RegionSelect to be one file, add test
